Advent of Code Day 2 Puzzle 2
This is basically the same code as Puzzle 1 with only a change in the
exit condition for the DO WHILE to find an MD5 hash with SIX leading
zeroes instead of five. It's not perfect it loops one extra time before
exiting. This ran a lot longer than the Puzzle 1 script but did get me
the correct answer.